scheduler/pyproject.toml
[tool.poetry]
name = "resaas-scheduler"
version = "0.1.0"
description = "Job scheduler for resaas"
authors = ["dorranh <dorran.howell@tweag.io>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^1.1.2"
apache-libcloud = "^3.3.0"
marshmallow = "^3.10.0"
SQLAlchemy = "^1.3.22"
psycopg2-binary = "^2.8.6"
paramiko = "^2.7.2"
Flask-SQLAlchemy = "^2.4.4"
flask-marshmallow = "^0.14.0"
marshmallow-sqlalchemy = "^0.24.1"
marshmallow-enum = "^1.5.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.1"
black = "^20.8b1"
flake8 = "^3.8.4"
pylint = "^2.6.0"
mypy = "^0.790"
libvirt-python = "^6.6.0"
pytest-dotenv = "^0.5.2"

[tool.black]
line-length = 99
target-version = ["py38"]
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
env_override_existing_values = 1
env_files = [".env"]
